data pipeline/get_data_bq_function/requirements.txt
pandas
google-cloud-storage
google-cloud-pubsub
google-cloud-bigquery